Remove filter form component
